ai-test-maker-backend/requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anyio==4.11.0
attrs==25.4.0
autoawq==0.2.9
bitsandbytes==0.48.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cmake==4.2.0
contourpy==1.3.3
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
datasets==4.4.1
dill==0.4.0
diskcache==5.6.3
filelock==3.20.0
fonttools==4.61.0
frozenlist==1.8.0
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
ImageIO==2.37.2
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
llama_cpp_python==0.3.16
lxml==6.0.2
Markdown==3.10
MarkupSafe==2.1.5
matplotlib==3.10.7
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
networkx==3.6
ninja==1.13.0
numpy==2.3.5
nvidia-cublas==13.0.0.19
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti==13.0.48
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc==13.0.48
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime==13.0.48
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cudnn-cu13==9.13.0.50
nvidia-cufft==12.0.0.15
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile==1.15.0.42
nvidia-curand==10.4.0.35
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver==12.0.3.29
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse==12.6.2.49
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu12==2.21.5
nvidia-nccl-cu13==2.27.7
nvidia-nvjitlink==13.0.39
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvshmem-cu13==3.3.24
nvidia-nvtx==13.0.39
nvidia-nvtx-cu12==12.4.127
packaging==25.0
pandas==2.3.3
pillow==12.0.0
propcache==0.4.1
psutil==7.1.3
pyarrow==22.0.0
PyMuPDF==1.26.6
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.10.16
tokenizers==0.21.4
torch==2.9.1+cu130
torchaudio==2.9.1+cu130
torchvision==0.24.1+cu130
tqdm==4.67.1
transformers==4.48.3
triton==3.5.1
typer-slim==0.20.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
